WIP editor addition.

+++ Source/Containers/CPU/OrganismInfoContainer.h
@@ -54,7 +54,6 @@ struct OrganismInfoContainer {
     }
 
 private:
-    //TODO possible optimization - iterative info updates.
     static void parse_organisms_info(OrganismInfoContainer & info, EngineDataContainer * edc, EngineControlParameters * ecp) {
         bool has_pool = true;
         int i = 0;
@@ -175,7 +174,7 @@ private:
         info.total_avg._killer_blocks     /= info.total_avg.total;
         info.total_avg._armor_blocks      /= info.total_avg.total;
         info.total_avg._eye_blocks        /= info.total_avg.total;
-        info.total_avg.occ_instructions_num = info.total_avg.total_occ_instructions_num > 0 ? info.total_avg.total_occ_instructions_num / info.total_avg.total : 0;
+        info.total_avg.occ_instructions_num = info.total_avg.total_occ_instructions_num > 0 ? (double)info.total_avg.total_occ_instructions_num / info.total_avg.total : 0;
 
         info.total_avg.brain_mutation_rate   /= info.total_avg.total;
         info.total_avg.anatomy_mutation_rate /= info.total_avg.total;
@@ -209,12 +208,12 @@ private:
         info.moving_avg._killer_blocks     /= info.moving_avg.total;
         info.moving_avg._armor_blocks      /= info.moving_avg.total;
         info.moving_avg._eye_blocks        /= info.moving_avg.total;
-        info.moving_avg.occ_instructions_num = info.moving_avg.total_occ_instructions_num > 0 ? info.moving_avg.total_occ_instructions_num / info.moving_avg.total : 0;
+        info.moving_avg.occ_instructions_num = info.moving_avg.total_occ_instructions_num > 0 ? (double)info.moving_avg.total_occ_instructions_num / info.moving_avg.total : 0;
 
         info.moving_avg.brain_mutation_rate   /= info.moving_avg.total;
         info.moving_avg.anatomy_mutation_rate /= info.moving_avg.total;
 
-        if (std::isnan(info.moving_avg.size))             {info.moving_avg.size             = 0;}
+        if (std::isnan(info.moving_avg.size))               {info.moving_avg.size               = 0;}
 
         if (std::isnan(info.moving_avg._organism_lifetime)) {info.moving_avg._organism_lifetime = 0;}
         if (std::isnan(info.moving_avg._organism_age))      {info.moving_avg._organism_age      = 0;}
@@ -242,12 +241,12 @@ private:
         info.station_avg._killer_blocks     /= info.station_avg.total;
         info.station_avg._armor_blocks      /= info.station_avg.total;
         info.station_avg._eye_blocks        /= info.station_avg.total;
-        info.station_avg.occ_instructions_num = info.station_avg.total_occ_instructions_num ? info.station_avg.total_occ_instructions_num / info.station_avg.total : 0;
+        info.station_avg.occ_instructions_num = info.station_avg.total_occ_instructions_num ? (double)info.station_avg.total_occ_instructions_num / info.station_avg.total : 0;
 
         info.station_avg.brain_mutation_rate   /= info.station_avg.total;
         info.station_avg.anatomy_mutation_rate /= info.station_avg.total;
 
-        if (std::isnan(info.station_avg.size))             {info.station_avg.size             = 0;}
+        if (std::isnan(info.station_avg.size))               {info.station_avg.size               = 0;}
 
         if (std::isnan(info.station_avg._organism_lifetime)) {info.station_avg._organism_lifetime = 0;}
         if (std::isnan(info.station_avg._organism_age))      {info.station_avg._organism_age      = 0;}
